Update and rename huffmancode.py to huffman_code.py
